Обновление от 19.07.2024

+++ backend/app/Http/Controllers/BlogController.php
@@ -12,28 +12,40 @@ use Symfony\Component\HttpFoundation\Response;
 class BlogController extends Controller
 {
     /**
-     * Получить все блоги с информацией об авторах.
+     * Список с авторами
+     * 
+     * @group Блоги
+     * @authenticated
      */
-    public function index()
+    public function index(Request $request)
     {
-        $blog = Blog::join('user_metadata', 'blogs.author_id', '=', 'user_metadata.user_id')
+        $perPage = $request->get('per_page', 10);
+        $blogs = Blog::join('user_metadata', 'blogs.author_id', '=', 'user_metadata.user_id')
             ->select('blogs.*', 'user_metadata.first_name', 'user_metadata.last_name', 'user_metadata.patronymic', 'user_metadata.nickname')
-            ->get();
+            ->paginate($perPage);
 
-        return response()->json($blog);
+        return response()->json($blogs);
     }
 
     /** 
-     * Получить блоги на основе предоставленных параметров запроса.
+     * Список
+     * 
+     * Получение списка блогов
+     * 
+     * @group Блоги
+     * 
+     * @bodyParam userId int ID пользователя.
+     * @bodyParam currentUser bool Флаг для поиска по текущему пользователю.
+     * @bodyParam blogId int ID блога.
+     * 
      */
     public function listBlogs(Request $request)
     {
-        // Параметры запроса: userId, currentUser 
-        // TODO: добавить имя автора
         $userId = $request->query('userId');
         $currentUser = $request->query('currentUser');
+        $blogId = $request->query('blogId');
         $response = [];
-        
+
 
         if ($userId) {
             $user = User::find($userId);
@@ -41,7 +53,7 @@ class BlogController extends Controller
             if (!$user) {
                 return $this->errorResponse('User not found', [], 404);
             }
-            
+
 
             $response[] = $user->blogs;
             $response[] = $user->metadata;
@@ -49,41 +61,147 @@ class BlogController extends Controller
             return $this->successResponse($response);
         } else if ($currentUser) {
             return $this->successResponse([Auth::user()->blogs, Auth::user()->metadata]);
+        } else if ($blogId) {
+            return $this->successResponse(Blog::find($blogId));
         } else {
-            
+
             return $this->successResponse(Blog::all(), '', 200);
         }
     }
 
 
     /**
-     * Show the form for creating a new resource.
+     * Список (новый)
+     * 
+     * Получение списка блогов (новый. использовать этот метод)
+     * 
+     * @group Блоги
+     * @authenticated
+     * 
+     * @bodyParam userId int ID пользователя.
+     * @bodyParam currentUser bool Флаг для поиска по текущему пользователю.
+     * @bodyParam blogId int ID блога.
+     * @urlParam withAuthors bool Включать авторов в ответ.
+     * @urlParam page int Номер страницы.
+     * @urlParam searchColumnName string Поиск по столбцу.
+     * @urlParam searchValue string Поисковый запрос.
+     * @urlParam tagFilter string Фильтр по тегу в meta описания.
+     * @urlParam crtFrom string Дата начала (формат: Y-m-d H:i:s).
+     * @urlParam crtTo string Дата окончания (формат: Y-m-d H:i:s).
+     * @urlParam updFrom string Дата начала (формат: Y-m-d H:i:s).
+     * @urlParam updTo string Дата окончания (формат: Y-m-d H:i:s).
+     * 
+     * @param \Illuminate\Http\Request $request
+     * @return mixed|\Illuminate\Http\JsonResponse
      */
-    public function create()
+    public function getBlogs(Request $request)
     {
-        //
+        if (!Auth::user()->can('view', Blog::class)) {
+            return $this->errorResponse('Нет прав на просмотр', [], 403);
+        }
+
+        $perPage = $request->get('per_page', 5);
+        $userId = $request->query('userId');
+        $currentUser = $request->query('currentUser');
+        $blogId = $request->query('blogId');
+        $withAuthors = $request->query('withAuthors', false);
+        $searchColumnName = $request->query('searchColumnName');
+        $searchValue = $request->query('searchValue');
+        $tagFilter = $request->query('tagFilter');
+        $crtFrom = $request->query('crtFrom');
+        $crtTo = $request->query('crtTo');
+        $updFrom = $request->query('updFrom');
+        $updTo = $request->query('updTo');
+
+        $query = Blog::query();
+
+        if ($withAuthors) {
+            $query->join('user_metadata', 'blogs.author_id', '=', 'user_metadata.user_id')
+                ->select('blogs.*', 'user_metadata.first_name', 'user_metadata.last_name', 'user_metadata.patronymic', 'user_metadata.nickname');
+        }
+
+        if ($userId) {
+            $user = User::find($userId);
+            if (!$user) {
+                return $this->errorResponse('User not found', [], 404);
+            }
+            $query->where('author_id', $userId);
+        } elseif ($currentUser) {
+            $currentUser = Auth::user();
+            if ($currentUser) {
+                $query->where('author_id', $currentUser->id);
+            } else {
+                return $this->errorResponse('Current user not found', [], 404);
+            }
+        } elseif ($blogId) {
+            $query->where('id', $blogId);
+        }
+
+        if ($searchColumnName) {
+            $query->where($searchColumnName, 'LIKE', '%' . $searchValue . '%');
+        }
+
+        if ($tagFilter) {
+            $query->whereRaw("description->'meta'->>'tags' LIKE ?", ['%' . $tagFilter . '%']);
+        }
+
+        if ($crtFrom && $crtTo) {
+            $query->whereBetween('created_at', [$crtFrom, $crtTo]);
+        } elseif ($crtFrom) {
+            $query->where('created_at', '>=', $crtFrom);
+        } elseif ($crtTo) {
+            $query->where('created_at', '<=', $crtTo);
+        }
+
+        if ($updFrom && $updTo) {
+            $query->whereBetween('updated_at', [$updFrom, $updTo]);
+        } elseif ($updFrom) {
+            $query->where('updated_at', '>=', $updFrom);
+        } elseif ($updTo) {
+            $query->where('updated_at', '<=', $updTo);
+        }
+
+        $blogs = $query->paginate($perPage);
+
+        $paginationData = [
+            'current_page' => $blogs->currentPage(),
+            'from' => $blogs->firstItem(),
+            'last_page' => $blogs->lastPage(),
+            'per_page' => $blogs->perPage(),
+            'to' => $blogs->lastItem(),
+            'total' => $blogs->total(),
+        ];
+
+        return $this->successResponse($blogs->items(), $paginationData, 200);
     }
 
+
+
+
     /**
-     * Store a newly created resource in storage.
+     * Создать
+     * 
+     * Создание нового блога
+     * 
+     * @group Блоги
+     * @authenticated
+     * 
      */
     public function store(Request $request)
     {
+        if (!Auth::user()->can('create')) {
+            return $this->errorResponse('Нет прав', [], 403);
+        }
+
         $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'required|string',
             'content' => 'required|string',
             'cover_uri' => 'nullable|string',
-            'status' => 'nullable|string|max:255',
-            'views' => 'nullable|integer',
-            'likes' => 'nullable|integer',
-            'reposts' => 'nullable|integer',
         ]);
 
         $input = $request->all();
-        Log::info($input);
         $input['status'] = 'moderating';
-        Log::info('awdaw', $input);
 
         // Создаем новый блог
         $blog = new Blog($input);
@@ -94,36 +212,36 @@ class BlogController extends Controller
         return response()->json(['message' => 'Blog created successfully', 'blog' => $blog], 201);
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(Blog $blog)
-    {
-        //
-    }
+
 
     /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Blog $blog)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
+     * Обновить
+     * 
+     * @authenticated
+     * @group Блоги
+     * 
+     * @bodyParam title string Название.
+     * @bodyParam description string Описание.
+     * @bodyParam content string Содержание.
+     * @bodyParam cover_uri string URI обложки.
+     * @bodyParam status string Статус.
+     * @bodyParam views int Количество просмотров.
+     * @bodyParam likes int Количество лайков.
+     * @bodyParam reposts int Количество репостов.
+     * 
      */
     public function update(Request $request, $id)
     {
-        $user = Auth::user();
-        if (!$user->hasPermissionTo('edit own news')) {
-            return $this->errorResponse('r', [], 403);
+        $blog = Blog::find($id);
+
+
+
+        if (!$blog) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
         }
 
-        $blogs = Blog::find($id);
-
-        if (!$blogs) {
-            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
+        if (!Auth::user()->can('update', $blog)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
         }
 
         $this->validateRequest($request, [
@@ -137,7 +255,6 @@ class BlogController extends Controller
             'reposts' => 'nullable|integer',
         ]);
 
-        // Фильтрация только тех полей, которые были переданы в запросе
         $updateData = $request->only([
             'title',
             'description',
@@ -149,25 +266,70 @@ class BlogController extends Controller
             'reposts',
         ]);
 
-        // Обновление данных новости
-        $blogs->update($updateData);
 
-        return $this->successResponse($blogs, 'Запись успешно обновлена', Response::HTTP_OK);
+
+        $blog->update($updateData);
+
+        return $this->successResponse($blog, 'Запись успешно обновлена', Response::HTTP_OK);
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Сменить статус
+     * 
+     * @group Блоги
+     * @authenticated
+     * 
+     * @bodyParam status string Новый статус
+     * 
+     */
+    public function setStatus(Request $request, $id)
+    {
+        $blog = Blog::find($id);
+
+        if (!$blog) {
+            return $this->errorResponse('Запись не найдена', [], Response::HTTP_NOT_FOUND);
+        }
+
+        if (!Auth::user()->can('changeStatus', $blog)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
+
+        $this->validateRequest($request, [
+            'status' => 'nullable|string|max:255',
+        ]);
+
+        $updateData = $request->only([
+            'status',
+        ]);
+
+        $blog->update($updateData);
+
+        return $this->successResponse($blog, 'Запись успешно обновлена', Response::HTTP_OK);
+    }
+
+    /**
+     * Удалить
+     * 
+     * @group Блоги
+     * @authenticated
+     * 
      */
     public function destroy($id)
     {
         $blog = Blog::find($id);
 
         if (!$blog) {
-            return response()->json(['error' => 'Blog not found'], Response::HTTP_NOT_FOUND);
+            return $this->errorResponse('Блог не найден', [], Response::HTTP_NOT_FOUND);
         }
+
+        if (!Auth::user()->can('delete', $blog)) {
+            return $this->errorResponse('Отсутствуют разрешения', [], 403);
+        }
+
+
 
         $blog->delete();
 
-        return response()->json(['success' => 'Blog successfully deleted'], Response::HTTP_OK);
+        return $this->successResponse(null, 'Запись удалена', Response::HTTP_OK);
     }
 }
